+++ src/components/Modal/index.tsx
@@ -16,7 +16,7 @@ export type ModalProps = {
   onToggle?: () => void;
   /**
    * Whether or not the top section with the close button should be absolute or not.
-   * If yes. It won't push the content down and the padding should be set on the children component.
+   * If yes. It won't push the content down and the hasPadding should be set on the children component.
    */
   isHeaderAbsolute?: boolean;
   /**
@@ -26,15 +26,15 @@ export type ModalProps = {
   /**
    * Whether or not close button should be displayed
    */
-  closeButton?: string;
+  isClosable?: string;
   /**
-   * Define a padding for content container
+   * What style of the componenteZ should be displayed
    */
-  paddingContent?: string;
+  variant?: 'Default' | 'New';
   /**
-   * Add css classes to content container
+   * Whether the content container should be removed
    */
-  containerClass?: string;
+  removePadding?: boolean;
   /**
    * A component that if provided will add a "actions" footer
    */
@@ -51,9 +51,9 @@ export type ModalProps = {
  * @param props.title A title string
  * @param props.actions A component that if provided will add a "actions" footer
  * @param props.children The provided children content
- * @param props.closeButton Whether or not close button should be displayed
- * @param props.containerClass Add css classes to content container
- * @param props.paddingContent Define a padding for content container
+ * @param props.isClosable Whether or not modal can be closed
+ * @param props.variant What style should be displayed
+ * @param props.removePadding Whether content container should be removed
  */
 export const Modal: React.FC<ModalProps> = ({
   onToggle,
@@ -61,9 +61,9 @@ export const Modal: React.FC<ModalProps> = ({
   isHeaderAbsolute,
   title,
   actions,
-  paddingContent = 'p-4',
-  containerClass,
-  closeButton = true,
+  removePadding,
+  variant = 'Default',
+  isClosable = true,
   children
 }) => {
   const [isOpen, setIsOpen] = useState(!!isOpenProps);
@@ -90,7 +90,7 @@ export const Modal: React.FC<ModalProps> = ({
      */
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const escape = (e: any) => {
-      if (e.keyCode === 27 && isOpen && onToggle) {
+      if (e.keyCode === 27 && isOpen && onToggle && isClosable) {
         onToggle();
       }
     };
@@ -100,22 +100,22 @@ export const Modal: React.FC<ModalProps> = ({
     return () => {
       document.removeEventListener('keydown', escape);
     };
-  }, [onToggle, isOpen]);
+  }, [onToggle, isOpen, isClosable]);
 
   return (
     <ReactPortal selector="#modal-root">
       <div className={classnames('pui-modal', { open: isOpen })}>
-        <div className="pui-modal-container">
+        <div className={classnames('pui-modal-container', variant)}>
           <div className={classnames('pui-modal-header ', { 'absolute z-30': isHeaderAbsolute })}>
-            <h3 className={paddingContent}>{title}</h3>
+            <h3 className={classnames(removePadding ? 'p-0' : 'p-4')}>{title}</h3>
             {/* Adds a close icon */}
-            {closeButton && (
-              <button className="pui-btn-icon text-pui-paragraph-900 px-6 py-4 " onClick={onToggle}>
+            {isClosable && (
+              <button className="pui-btn-icon text-pui-paragraph-900 px-6 py-4" onClick={onToggle}>
                 <CrossIcon className="pui-animate-scaleHover-target" />
               </button>
             )}
           </div>
-          <div className={classnames('pui-modal-content', containerClass, paddingContent)}>{children}</div>
+          <div className={classnames('pui-modal-content', removePadding ? 'p-0' : 'p-4')}>{children}</div>
           {actions && <div className="pui-modal-actions">{actions()}</div>}
         </div>
       </div>